Refactor speech functionality
